Re-arrange searchbar logic & structure into its own component for better organization (#15)
